# gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Policy Service
    description: The policy service exposes HTTP API for executing policies.
    version: "1.0"
servers:
    - url: http://localhost:8081
      description: Policy Server
paths:
...
    /policy/{repository}/{group}/{policyName}/{version}/evaluation/did.json:
        get:
            tags:
                - policy
            summary: Evaluate policy
            description: Evaluate executes a policy with the given 'data' as input.
            operationId: policy#Evaluate
            parameters:
                - name: repository
                  in: path
                  description: Policy repository.
                  required: true
                  schema:
                    type: string
                    description: Policy repository.
                    example: policies
                  example: policies
                - name: group
                  in: path
                  description: Policy group.
                  required: true
                  schema:
                    type: string
                    description: Policy group.
                    example: example
                  example: example
                - name: policyName
                  in: path
                  description: Policy name.
                  required: true
                  schema:
                    type: string
                    description: Policy name.
                    example: example
                  example: example
                - name: version
                  in: path
                  description: Policy version.
                  required: true
                  schema:
                    type: string
                    description: Policy version.
                    example: "1.0"
                  example: "1.0"
                - name: x-evaluation-id
                  in: header
                  description: EvaluationID allows overwriting the randomly generated evaluationID
                  allowEmptyValue: true
                  schema:
                    type: string
                    description: EvaluationID allows overwriting the randomly generated evaluationID
                    example: did:web:example.com
                  example: did:web:example.com
                - name: x-cache-ttl
                  in: header
                  description: Policy result cache TTL in seconds
                  allowEmptyValue: true
                  schema:
                    type: integer
                    description: Policy result cache TTL in seconds
                    example: 60
                    format: int64
                  example: 60
            requestBody:
                description: Input data passed to the policy execution runtime.
                required: true
                content:
                    application/json:
                        schema:
                            type: string
                            description: Input data passed to the policy execution runtime.
                            example: Et numquam non rerum.
                            format: binary
                        example: Aut doloremque beatae non sed nihil perferendis.
            responses:
                "200":
                    description: OK response.
                    headers:
                        ETag:
                            description: ETag contains unique identifier of the policy evaluation and can be used to later retrieve the results from Cache.
                            schema:
                                type: string
                                description: ETag contains unique identifier of the policy evaluation and can be used to later retrieve the results from Cache.
                                example: Id distinctio perspiciatis.
                            example: Consequatur cupiditate aut consequuntur in animi.
                    content:
                        application/json:
                            schema:
                                type: string
                                description: Arbitrary JSON response.
                                example: Quis eius voluptas est ipsum.
                                format: binary
                            example: Aspernatur ut ab nam quis repellendus.
    /policy/{repository}/{group}/{policyName}/{version}/export:
        get:
            tags:
                - policy
            summary: ExportBundle policy
            description: Export a signed policy bundle.
            operationId: policy#ExportBundle
            parameters:
                - name: repository
                  in: path
                  description: Policy repository.
                  required: true
                  schema:
                    type: string
                    description: Policy repository.
                    example: policies
                  example: policies
                - name: group
                  in: path
                  description: Policy group.
                  required: true
                  schema:
                    type: string
                    description: Policy group.
                    example: example
                  example: example
                - name: policyName
                  in: path
                  description: Policy name.
                  required: true
                  schema:
                    type: string
                    description: Policy name.
                    example: returnDID
                  example: returnDID
                - name: version
                  in: path
                  description: Policy version.
                  required: true
                  schema:
                    type: string
                    description: Policy version.
                    example: "1.0"
                  example: "1.0"
            responses:
                "200":
                    description: OK response.
                    headers:
                        content-disposition:
                            description: Content-Disposition response header containing the name of the file.
                            schema:
                                type: string
                                description: Content-Disposition response header containing the name of the file.
                                example: Fuga ex vero corporis non.
                            example: Illum porro mollitia ducimus assumenda rerum.
                        content-length:
                            description: Content-Length response header.
                            schema:
                                type: integer
                                description: Content-Length response header.
                                example: 8893876685428017835
                                format: int64
                            example: 2727357281425983130
                        content-type:
                            description: Content-Type response header.
                            schema:
                                type: string
                                description: Content-Type response header.
                                example: Dolores iusto corporis quos recusandae.
                            example: Fugit non incidunt ut quidem doloremque.
                    content:
                        application/json:
                            schema:
                                type: string
                                format: binary
...
    /policy/{repository}/{group}/{policyName}/{version}/notifychange:
        post:
            tags:
                - policy
            summary: SubscribeForPolicyChange policy
            description: Subscribe for policy change notifications by registering webhook callbacks which the policy service will call.
            operationId: policy#SubscribeForPolicyChange
            parameters:
                - name: repository
                  in: path
                  description: Policy repository.
                  required: true
                  schema:
                    type: string
                    description: Policy repository.
                    example: Ducimus ut non molestiae veniam aut est.
                  example: Ut perferendis.
                - name: group
                  in: path
                  description: Policy group.
                  required: true
                  schema:
                    type: string
                    description: Policy group.
                    example: Quia sed et quis fugit ipsam tempora.
                  example: Nobis officiis natus illo ex in.
                - name: policyName
                  in: path
                  description: Policy name.
                  required: true
                  schema:
                    type: string
                    description: Policy name.
                    example: In ab sed excepturi.
                  example: Aut vero quidem non et ut nihil.
                - name: version
                  in: path
                  description: Policy version.
                  required: true
                  schema:
                    type: string
                    description: Policy version.
                    example: Consequuntur sunt autem est ipsa veritatis hic.
                  example: Aperiam nihil sint nostrum.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SubscribeForPolicyChangeRequestBody'
                        example:
                            subscriber: nn6
                            webhook_url: http://white.com/grover
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: string
                                example: Facere sint ipsum saepe ut.
                                format: binary
                            example: Autem aut et recusandae et.
...
    /v1/policies:
        get:
            tags:
                - policy
            summary: ListPolicies policy
            description: List policies from storage with optional filters.
            operationId: policy#ListPolicies
            parameters:
                - name: locked
                  in: query
                  description: Filter to return locked/unlocked policies (optional).
                  allowEmptyValue: true
                  schema:
                    type: boolean
                    description: Filter to return locked/unlocked policies (optional).
                    example: false
                  example: true
                - name: rego
                  in: query
                  description: Include policy source code in results (optional).
                  allowEmptyValue: true
                  schema:
                    type: boolean
                    description: Include policy source code in results (optional).
                    example: false
                  example: true
                - name: data
                  in: query
                  description: 'Include policy static data in results (optional). '
                  allowEmptyValue: true
                  schema:
                    type: boolean
                    description: 'Include policy static data in results (optional). '
                    example: true
                  example: false
                - name: dataConfig
                  in: query
                  description: Include static data config (optional).
                  allowEmptyValue: true
                  schema:
                    type: boolean
                    description: Include static data config (optional).
                    example: true
                  example: false
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PoliciesResult'
                            example:
                                policies:
                                    - data: Similique architecto.
                                      dataConfig: Sunt ut est ut molestias.
                                      group: Error minus vel voluptate quasi.
                                      lastUpdate: 2308837710267791288
                                      locked: true
                                      policyName: Vel dolores omnis molestiae tempora sed repellendus.
                                      rego: Vel sunt dolorem ea architecto iure.
                                      repository: Ut iure aut asperiores.
                                      version: Debitis consectetur.
                                    - data: Similique architecto.
                                      dataConfig: Sunt ut est ut molestias.
                                      group: Error minus vel voluptate quasi.
                                      lastUpdate: 2308837710267791288
                                      locked: true
                                      policyName: Vel dolores omnis molestiae tempora sed repellendus.
                                      rego: Vel sunt dolorem ea architecto iure.
                                      repository: Ut iure aut asperiores.
                                      version: Debitis consectetur.
                                    - data: Similique architecto.
                                      dataConfig: Sunt ut est ut molestias.
                                      group: Error minus vel voluptate quasi.
                                      lastUpdate: 2308837710267791288
                                      locked: true
                                      policyName: Vel dolores omnis molestiae tempora sed repellendus.
                                      rego: Vel sunt dolorem ea architecto iure.
                                      repository: Ut iure aut asperiores.
                                      version: Debitis consectetur.
components:
    schemas:
...
        PoliciesResult:
            type: object
            properties:
                policies:
                    type: array
                    items:
                        $ref: '#/components/schemas/Policy'
                    description: JSON array of policies.
                    example:
                        - data: Similique architecto.
                          dataConfig: Sunt ut est ut molestias.
                          group: Error minus vel voluptate quasi.
                          lastUpdate: 2308837710267791288
                          locked: true
                          policyName: Vel dolores omnis molestiae tempora sed repellendus.
                          rego: Vel sunt dolorem ea architecto iure.
                          repository: Ut iure aut asperiores.
                          version: Debitis consectetur.
                        - data: Similique architecto.
                          dataConfig: Sunt ut est ut molestias.
                          group: Error minus vel voluptate quasi.
                          lastUpdate: 2308837710267791288
                          locked: true
                          policyName: Vel dolores omnis molestiae tempora sed repellendus.
                          rego: Vel sunt dolorem ea architecto iure.
                          repository: Ut iure aut asperiores.
                          version: Debitis consectetur.
            example:
                policies:
                    - data: Similique architecto.
                      dataConfig: Sunt ut est ut molestias.
                      group: Error minus vel voluptate quasi.
                      lastUpdate: 2308837710267791288
                      locked: true
                      policyName: Vel dolores omnis molestiae tempora sed repellendus.
                      rego: Vel sunt dolorem ea architecto iure.
                      repository: Ut iure aut asperiores.
                      version: Debitis consectetur.
                    - data: Similique architecto.
                      dataConfig: Sunt ut est ut molestias.
                      group: Error minus vel voluptate quasi.
                      lastUpdate: 2308837710267791288
                      locked: true
                      policyName: Vel dolores omnis molestiae tempora sed repellendus.
                      rego: Vel sunt dolorem ea architecto iure.
                      repository: Ut iure aut asperiores.
                      version: Debitis consectetur.
            required:
                - policies
        Policy:
            type: object
            properties:
                data:
                    type: string
                    description: Policy static data.
                    example: Qui et sit maiores architecto alias.
                dataConfig:
                    type: string
                    description: Policy static data optional configuration.
                    example: Nesciunt labore voluptatibus.
                group:
                    type: string
                    description: Policy group.
                    example: Libero sed a at.
                lastUpdate:
                    type: integer
                    description: Last update (Unix timestamp).
                    example: 731402412785725769
                    format: int64
                locked:
                    type: boolean
                    description: Locked specifies if the policy is locked or allowed to execute.
                    example: false
                policyName:
                    type: string
                    description: Policy name.
                    example: Deleniti odit dolor et et.
                rego:
                    type: string
                    description: Policy rego source code.
                    example: Quia blanditiis.
                repository:
                    type: string
                    description: Policy repository.
                    example: Rerum exercitationem odit tempora ab in aliquid.
                version:
                    type: string
                    description: Policy version.
                    example: Qui delectus.
            example:
                data: Debitis nam voluptas qui quisquam.
                dataConfig: Aut voluptatem consequatur totam reiciendis molestiae itaque.
                group: Porro rerum.
                lastUpdate: 5292309736494809727
                locked: false
                policyName: Autem quos autem aut in est dolor.
                rego: Provident corrupti rerum exercitationem placeat.
                repository: Deserunt expedita facilis.
                version: Tempore quis qui.
            required:
                - repository
                - group
                - policyName
                - version
                - locked
                - lastUpdate
        SubscribeForPolicyChangeRequestBody:
            type: object
            properties:
                subscriber:
                    type: string
                    description: Name of the subscriber for policy.
                    example: tkm
                    minLength: 3
                    maxLength: 100
                webhook_url:
                    type: string
                    description: Subscriber webhook url.
                    example: http://dickensgrady.name/kelvin
                    format: uri
            example:
                subscriber: 8yb
                webhook_url: http://ziemann.name/loraine.wilderman
            required:
                - webhook_url
                - subscriber
...